Handle signals in consumer rake task.
When upstart stops the worker, it sends a SIGTERM to the process.  This
is manifested in ruby as a SignalException.  Unless handled, this
exception bubbles all the way up, and it intercepted by the airbrake
handler, resulting in noise in Errbit.

Handling this exception in the rake task prevents these spurious
exception reports in Errbit every time the workers are restarted.