Update and configure ci_reporter.
This was only partially installed.  Also needed updating to avoid
deprecation warnings with RSpec3

It was necessary to disable its output capture feature.  This replaces
STDOUT and STDERR with a StringIO.  This causes issues with childprocess
which expects STDOUT and STDERR to have a file descriptor so that it can
be dup'd into the child.