Expand out "links" on Entry JSON representation
We want entries to include the expanded form of their links. Eventually, these
will include details such as the title, format, base_path, etc, but that is a
much larger piece of work that requires de-referencing content_ids against other
entries in content-register. For now, it is enough for us to partially expand
the links, only including the "content_id". This is a future-proof
data-structure that can be expanded later to include the other fields.