Ensure default column limit is removed
For some reason this had previously worked, I was unable to track down what
may have changed in rails 4.1.11 (since 4.1.5).

The database tasks used to depend on the rails env which would have triggered this monkey
patch to be loaded, but that dependency was removed in rails 4.0.0[1], which implies that
tests should have been failing already.

[1] https://github.com/rails/rails/commit/f1afd7765c2a35cc77e1ee40466db36f181e29f5