Fix handling of database uniqueness violations.
When a database uniqueness constraint is violated, ActiveRecord wraps
the exception from the database driver in an
ActiveRecord::RecordNotUnique. This wasn't obvious because Airbrake
unwraps this exception before sending the error to Errbit [1].

This updates the rescue clause in the message queue processor to rescue
the correct exception class, and therefore handle the errors.

[1] https://github.com/airbrake/airbrake/blob/master/lib/airbrake.rb#L181-L187