Strip non-UUID links on receipt.
We're sending some expanded links onto the queue
which breaches this app's validations.

Since these expanded links aren't used by the apps
which use content register, the simplest solution
is to strip them out on receipt.

This may need to be revisited when we begin sending
only expanded links onto the queue, but hopefully
this app will not exist by then.